requirements.txt
rasterio
ultralytics
opencv-python-headless
pandas
openpyxl
git+https://github.com/facebookresearch/segment-anything.git
jupyter_bbox_widget
roboflow
dataclasses-json
supervision==0.23.0
scikit-image
scikit-learn
joblib
torch
streamlit>=1.24.0
streamlit-extras
numpy>=1.21.0
Pillow>=8.3.1
shap>=0.42.1
matplotlib>=3.4.0
langchain-openai
tiktoken
PyYAML
